backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-jose==3.3.0
python-multipart==0.0.9
werkzeug==3.0.3
pydantic[email]==2.5.3
pydantic-settings==2.0.3

